+++ config/Location.hpp
@@ -5,8 +5,8 @@
 #include <set>
 #include <string>
 
-#include "../util/Config.hpp"
-#include "../util/Enum.hpp"
+#include "../utils/Config.hpp"
+#include "../utils/Enum.hpp"
 
 // config 파일의 Location 블럭
 // - location 블럭의 정보를 저장한다.
